backend/rag_chatbot/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
qdrant-client==1.7.0
openai==1.12.0
python-dotenv==1.0.0
markdown==3.5.1
langchain==0.0.339
tiktoken==0.5.2
typing-extensions>=4.7.0